requirements.txt
# Add project dependencies here 
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
fhir.resources>=6.4.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
python-dotenv>=0.19.0
asyncpg>=0.27.0
greenlet>=3.2.1  